:safety_vest: Add jwt guard on users endpoints

+++ src/modules/users/users.controller.ts
@@ -1,7 +1,7 @@
 import {
   Controller,
   Get,
-  Post,
+  Request,
   Body,
   Patch,
   Param,
@@ -9,11 +9,13 @@ import {
   HttpCode,
   HttpStatus,
   ParseUUIDPipe,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
+import { AuthGuard } from '@nestjs/passport';
 
 @Controller('users')
 export class UsersController {
@@ -24,12 +26,14 @@ export class UsersController {
   }
 
   @Get()
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(HttpStatus.OK)
   findAll(): Promise<User[]> {
     return this.usersService.findAll();
   }
 
   @Get(':id')
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(HttpStatus.OK)
   async findOne(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
@@ -38,19 +42,33 @@ export class UsersController {
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(HttpStatus.OK)
   update(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() updateUserDto: UpdateUserDto,
+    @Request() request,
   ): Promise<User> {
+    const tokenUserId = request.user.id;
+
+    if (tokenUserId !== id)
+      throw new ForbiddenException('User can not update a different user');
+
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(HttpStatus.OK)
   remove(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
-  ): Promise<User> {
+    @Request() request,
+  ): Promise<User> | string {
+    const tokenUserId = request.user.id;
+
+    if (tokenUserId !== id)
+      throw new ForbiddenException('User can not delete a different user');
+
     return this.usersService.remove(id);
   }
 }
